v1.3.1
You'd be surprised how popular nametag removers are. So we added an option to serve as one if that's what people really want. Ours even comes with the benefit of changing the tab list whilst removing above the head names :D In the future we should allow some more flexibility around the visible tag to cover the options available in game. 

- Fixed a bug where weight was not considered for applying nametags. 
- Added the new, optional `visible: <true/false>` option to tags.